requirements.txt
python-dotenv
google-genai
pyperclip
streamlit
watchdog